-rename file to load.py - move loadRadianceObj to load.py
