+++ bifacial_radiance/bifacial_radiance.py
@@ -135,33 +135,6 @@ def _interactive_directory(title=None):
     root.attributes("-topmost", True) #Bring to front
     return filedialog.askdirectory(parent = root, title = title)
 
-def loadRadianceObj(savefile=None):
-        '''
-        Load the pickled radiance object for further use
-        usage: (once you're in the correct local directory)
-          demo = bifacial_radiance.loadRadianceObj(savefile)
-        
-        Parameters
-        ----------
-        savefile :   optional savefile.  Otherwise default to save.pickle
-                
-        '''
-        import pickle
-        
-        if savefile is None:
-            savefile = 'save.pickle'
-        with open(savefile,'rb') as f:
-            loadObj= pickle.load(f)
-        
-        print('Loaded file {}'.format(savefile))
-        return loadObj
-
-
-        
-      
-
-
-
  
 
 class RadianceObj:
